Part Two
An abstract class (Person) and an interface (RideInterface) are used.

The Ride class implements all the methods in the interface, managing guest queues and ride history.

The test class simulates the operation of the theme park, including adding visitors, running rides, removing visitors, printing information, etc. (This test is not reflected in the code.)