Part Four
A: In the Ride class, I changed the rewriting method of the RideInterface interface. In the runOneCycle() method, I used the poll method to throw out the contents of the queue based on the first-in-first-out principle, and added the thrown tourists to the rideHistory list. In the checkVisitorFromHistory method, I use a ternary operator on the output. In the printRideHistory() method, it's changed to iterators that iterate over the list.

B: I wrote a class that interfaces with Comparator<Visitor>, which will be used to compare Visitor objects. I ranked the age of the visitors from youngest to oldest, if the same, this compares the names. Then, in the Ride class, compare with the sort method.