Part Five
I added two properties to the Ride class, the maxRider property and the numOfCycles property, and I set the maxRider property to public because I wanted to assign it in the test class. I also implemented the RunOneCycle method so that it runs until there are no tourists waiting in the queue. I also completed part5 of the test class, where I assigned the instance object's maxRider property to 4 so that it could run three times.