Part VI
I have added the exportRideHistory method in order to export the history of visitors' access to entertainment facilities into a given text file. The details of each visitor are written to the file in CSV (comma-separated values) format.

I used the character buffered output stream BufferedWriter to wrap FileWriter to write characters, arrays, and strings more efficiently.

I also use the try-with-resources statement to ensure that the writer is automatically closed when the block is exited, thus preventing resource leakage.

Second, if an IOException occurs (for example, a file cannot be created or written), the error message is printed.