Part 7
In this part, I created importRideHistory, and as in the previous part, I also used the character buffered output stream BufferedReader to read the file line by line, and I also used the try-with-resources statement to ensure that the reader is automatically closed after use. To avoid resource leakage.
When reading the file, on each line I separate the data into an array of strings based on commas. The code checks that the separated array contains exactly 10 elements to ensure that the data is in the correct format before processing it. If the line does not contain exactly 10 elements, an error message is printed indicating that the data format is invalid.
If the data is in the correct format, a new Visitor object is created using the parsed data. Each element in the data array corresponds to an attribute of the Visitor, where age is parsed as an integer.
The IOException handles any ioExceptions that may occur while reading the file and prints out an error message indicating the nature of the problem.